pyproject.toml
[project]
name = "langchain-chapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
]
